requirements.txt
flask==3.0.2
flask-cors==4.0.0
torch==2.1.2
torchvision==0.16.2
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0

# YOLOv5 dependencies
matplotlib>=3.3.0
pyyaml>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
pandas>=1.1.4
seaborn>=0.11.0
psutil>=5.9.5

# Vision AI (optional)
groq==0.4.0
python-dotenv==1.0.1

# Development tools
black==24.2.0
pylint==3.0.3
pytest==8.0.2

# GPU support (comment out if not using CUDA)
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.2+cu118 
torchvision==0.16.2+cu118